Move attrs_plugin to plugins/attrs
